Adjust mbox parsing to search for subcomponents' names
Search for the names of subcomponents in message contents.

In this commit, some code statements are slightly changed for that
reason.

Important: This is now independent from any codeface configuration. You
can just specify the path to the mbox file, the path to the commits list
from which to extract the subcomponents' names, and the path where to
store the result file.

Example call:

python run-parsing.py -p test ${OUTPUT_PATH} ${PATH_TO_MBOX_FILE} ${PATH_TO_COMMITS_LIST}

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>