Adjust mbox parsing to search for subcomponents' names
Search not only for complete occurrences in the index, that is, complete
words resulting from the tokenization, but also for partly occurrences.

Example: If we search for "Documentation/RCU/Desgin", formerly, we did
not get a match if
"Documentation/RCU/Design/Requirements/Requirements.htmlx" occurs in an
e-mail, as there is no token in the index which exactly matches the
search query.

To fix this, we use a regular expression to search for any tokens which
begin with the search query. Therefore, we need to import the
'RegexPlugin' from the whoosh query parser and adjust the query.

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>