Fix regex for subcomponent search again
The regex for the subcomponent search did not match in some cases for
the following three reasons:

(1) The diff in patches needs not to follow the pattern "a/file b/file",
as "a" or "b" can be replaced by something else. In particular, the
following patterns do exist:

- "old/file" and "new/file"
- "linux-2.6-git/file" and "linux-2.6-git_xxx/file"
- "linux-2.6-git~/file" and "linux-2.6-git+/file"
- "linux/file" and "./file"
- "l1/file" and "l2/file"
- "linus/file" and "linus/file~"
- "{arbitrary alpha-numeric string}/file"
   and "{arbitrary alpha-numeric string}.orig/file"
- "{arbitrary alpha-numberic string}.patch-new/file"
   and "{arbitrary alpha-numeric string}.patch-old/file"
- "x-0/file" and "x-1/file"

This is fixed by adding "(linux\-[\w\-\.:+~]*)", "\.", "linux", "linus",
"old", "new", "l1, "l2", "x-0", "x-1", and
"([\w\-\.]*\.(patch\-old|patch\-new|orig))" as additional alternatives
to "a" and "b".

However, there are lots of different ways how the files of the diff
can be named (e.g., individual directory names at the beginning). As
those file paths cannot be mapped to a specific pattern, subcomponent
names in such file paths cannot be found by our regex.

(2) Many file names containing subcomponent names could not be matched
due to an obvious reason: "." and "-" were not contained in the group of
allowed characters occurring after the subcomponent name, e.g., in file
extensions. This is fixed by allowing "." and "-" to appear after the /
after the subcomponent name.

(3) Many file names containing subcomponent names could not be matched
if there were additional sub-directories in-between subcomponent name
and the actual file name. E.g., "subcomponent/s1/s2/file.c" could not
be matched as the slashes after s1 and s2 would not be allowed.
Therefore, the "/" is added to the allowed characters appearing after
the / after the subcomponent name as well.

With these three fixes, the regex for subcomponent search should be
valid now to match (almost) all expected cases.

Props to @clhunsen for supporting me in fixing part (1).

Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>